대쉬보드 요구사항 수정

+++ smart-web/src/main/java/net/woori/start/domain/weather/WeatherInfo.java
@@ -17,7 +17,7 @@ public class WeatherInfo {
 	private String type;
 	
 	/** 강우량 / 강수확률 */
-	private int rainfall;
+	private float rainfall;
 	
 	/** 미세먼지 */
 	private Status fineDust;
